SBERDOMA-1122 moved `addressMeta` helpers and factories to "property" domain
