+++ apps/condo/domains/property/schema/CheckPropertyWithAddressExistService.test.js
@@ -5,7 +5,7 @@ const { makeClientWithProperty } = require('@condo/domains/property/utils/testSc
 const { checkPropertyWithAddressExistByTestClient } = require('@condo/domains/property/utils/testSchema')
 const { catchErrorFrom } = require('@condo/domains/common/utils/testSchema')
 const { DV_VERSION_MISMATCH_MESSAGE } = require('./CheckPropertyWithAddressExistService')
-const { FLAT_WITHOUT_FLAT_TYPE_MESSAGE } = require('@condo/domains/common/utils/addressApi/helpers')
+const { FLAT_WITHOUT_FLAT_TYPE_MESSAGE } = require('../utils/serverSchema/helpers')
 
 describe('CheckPropertyWithAddressExistService', async () => {
     test('user finds result', async () => {
