+++ apps/condo/domains/property/utils/serverSchema/helpers.js
