+++ apps/condo/domains/property/utils/testSchema/factories.js
@@ -1,6 +1,6 @@
 const faker = require('faker')
 const { sample } = require('lodash')
-const { AddressMetaDataFields } = require('../addressApi/AddressMetaSchema')
+const { AddressMetaDataFields } = require('../../schema/fields/AddressMetaField')
 
 const FIAS_LEVELS = [1, 2, 3, 4, 5, 6, 7, 8, 9, 35, 65, 75, 90, 91]
 const FIAS_ACTUALITY_STATE = [0, 1]
