+++ bin/generate-receipts.js
@@ -7,7 +7,7 @@
 const { BillingProperty, BillingAccount, BillingReceipt, BillingIntegrationOrganizationContext } = require('@condo/domains/billing/utils/serverSchema')
 const faker = require('faker')
 const path = require('path')
-const { buildFakeAddressMeta } = require('@condo/domains/common/utils/testSchema/factories')
+const { buildFakeAddressMeta } = require('@condo/domains/property/utils/testSchema/factories')
 const { GraphQLApp } = require('@keystonejs/app-graphql')
 const { getPreviousPeriods } = require('@condo/domains/billing/utils/period')
 const { DateTime } = require('luxon')
